requirements.txt
pyyaml==6.0.1
eyeD3==0.9.7
requests==2.31.0
pillow==10.2.0